add .c and .h files
